cmd/snap-update-ns: move secureOpenPath implementation to utils.go
